refact: move display components to new subfolder
